Fixing neighbors.size()

+++ inst/include/epiworld/agent-meat.hpp
@@ -690,7 +690,7 @@ inline void Agent<TSeq>::print(
             "Agent: %i, state: %s (%i), Has virus: %s, NTools: %i, NNeigh: %i\n",
             id, model->states_labels[state].c_str(), static_cast<int>(state),
             virus == nullptr ? std::string("no").c_str() : std::string("yes").c_str(),
-            static_cast<int>(n_tools), neighbors.size()
+            static_cast<int>(n_tools), static_cast<int>(neighbors.size())
         );
     }
     else {
@@ -700,7 +700,7 @@ inline void Agent<TSeq>::print(
         printf_epiworld("  Has virus    : %s\n", virus == nullptr ?
             std::string("no").c_str() : std::string("yes").c_str());
         printf_epiworld("  Tool count   : %i\n", static_cast<int>(n_tools));
-        printf_epiworld("  Neigh. count : %i\n", neighbors.size());
+        printf_epiworld("  Neigh. count : %i\n", static_cast<int>(neighbors.size()));
 
         size_t nfeats = model->get_agents_data_ncols();
         if (nfeats > 0)
